+++ jolt-core/src/zkvm/lookup_table/or.rs
@@ -11,9 +11,9 @@ use crate::utils::uninterleave_bits;
 pub struct OrTable<const WORD_SIZE: usize>;
 
 impl<const WORD_SIZE: usize> JoltLookupTable for OrTable<WORD_SIZE> {
-    fn materialize_entry(&self, index: u64) -> u64 {
+    fn materialize_entry(&self, index: u128) -> u64 {
         let (x, y) = uninterleave_bits(index);
-        (x | y) as u64
+        x | y
     }
 
     fn evaluate_mle<F: JoltField>(&self, r: &[F]) -> F {
