+++ jolt-core/src/zkvm/lookup_table/signed_greater_than_equal.rs
@@ -11,7 +11,7 @@ use crate::{field::JoltField, utils::uninterleave_bits};
 pub struct SignedGreaterThanEqualTable<const WORD_SIZE: usize>;
 
 impl<const WORD_SIZE: usize> JoltLookupTable for SignedGreaterThanEqualTable<WORD_SIZE> {
-    fn materialize_entry(&self, index: u64) -> u64 {
+    fn materialize_entry(&self, index: u128) -> u64 {
         let (x, y) = uninterleave_bits(index);
         match WORD_SIZE {
             #[cfg(test)]
