+++ jolt-core/src/zkvm/lookup_table/suffixes/lower_word.rs
@@ -7,7 +7,7 @@ use super::SparseDenseSuffix;
 pub enum LowerWordSuffix<const WORD_SIZE: usize> {}
 
 impl<const WORD_SIZE: usize> SparseDenseSuffix for LowerWordSuffix<WORD_SIZE> {
-    fn suffix_mle(b: LookupBits) -> u32 {
-        (u64::from(b) % (1 << WORD_SIZE)) as u32
+    fn suffix_mle(b: LookupBits) -> u64 {
+        (u128::from(b) % (1 << WORD_SIZE)) as u64
     }
 }
