+++ jolt-core/src/zkvm/lookup_table/suffixes/upper_word.rs
@@ -7,7 +7,7 @@ use super::SparseDenseSuffix;
 pub enum UpperWordSuffix<const WORD_SIZE: usize> {}
 
 impl<const WORD_SIZE: usize> SparseDenseSuffix for UpperWordSuffix<WORD_SIZE> {
-    fn suffix_mle(b: LookupBits) -> u32 {
-        (u64::from(b) >> WORD_SIZE) as u32
+    fn suffix_mle(b: LookupBits) -> u64 {
+        (u128::from(b) >> WORD_SIZE) as u64
     }
 }
