+++ jolt-core/src/zkvm/lookup_table/unsigned_less_than.rs
@@ -11,7 +11,7 @@ use super::{
 pub struct UnsignedLessThanTable<const WORD_SIZE: usize>;
 
 impl<const WORD_SIZE: usize> JoltLookupTable for UnsignedLessThanTable<WORD_SIZE> {
-    fn materialize_entry(&self, index: u64) -> u64 {
+    fn materialize_entry(&self, index: u128) -> u64 {
         let (x, y) = uninterleave_bits(index);
         (x < y).into()
     }
