+++ jolt-core/src/zkvm/r1cs/key.rs
@@ -203,7 +203,7 @@ impl<F: JoltField> UniformSpartanKey<F> {
         // Evaluate at the constant position if it exists within the padded space
         let const_eval = if self.uniform_r1cs.num_vars < num_vars && with_const {
             let const_position_bits =
-                index_to_field_bitvector(self.uniform_r1cs.num_vars as u64, var_bits);
+                index_to_field_bitvector(self.uniform_r1cs.num_vars as u128, var_bits);
             EqPolynomial::mle(r, &const_position_bits)
         } else {
             F::zero()
@@ -238,8 +238,8 @@ impl<F: JoltField> UniformSpartanKey<F> {
 
         // Evaluate non-constant terms
         for (row, col, val) in constraints.vars.iter() {
-            let row_bits = index_to_field_bitvector(*row as u64, rx_constr.len());
-            let col_bits = index_to_field_bitvector(*col as u64, ry_var.len());
+            let row_bits = index_to_field_bitvector(*row as u128, rx_constr.len());
+            let col_bits = index_to_field_bitvector(*col as u128, ry_var.len());
             eval += *val
                 * EqPolynomial::mle(rx_constr, &row_bits)
                 * EqPolynomial::mle(ry_var, &col_bits);
@@ -247,11 +247,11 @@ impl<F: JoltField> UniformSpartanKey<F> {
 
         // Evaluate constant terms
         let constant_column = self.uniform_r1cs.num_vars;
-        let const_col_bits = index_to_field_bitvector(constant_column as u64, ry_var.len());
+        let const_col_bits = index_to_field_bitvector(constant_column as u128, ry_var.len());
         let eq_ry_const = EqPolynomial::mle(ry_var, &const_col_bits);
 
         for (row, val) in constraints.consts.iter() {
-            let row_bits = index_to_field_bitvector(*row as u64, rx_constr.len());
+            let row_bits = index_to_field_bitvector(*row as u128, rx_constr.len());
             eval += *val * EqPolynomial::mle(rx_constr, &row_bits) * eq_ry_const;
         }
